Content, components and somantics
Changed h2 > p tags in the heading of the page as it is somantically more correct. Added components to several pages and started filling out content. Updated table layout to prevent them being too wide on the page. Added a heading component for headings on pages. Sorted out some of the parent directories for some of the pages